refactor(chapter08): rename case02.py to case02_1.py for clarity in module structure
